+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bump clj-londo 2021.03.22-20210327.192113-4 - Fixes #385
 - Add support for outgoing call hierarchy - Fixes #384
+- Improve and fix missing completion item kinds.
 
 ## 2021.03.26-23.41.07