+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2021.03.30-20.42.34
+
 - Bump clj-londo 2021.03.22-20210327.192113-4 - Fixes #385
 - Add support for outgoing call hierarchy - Fixes #384
 - Improve and fix missing completion item kinds.